Put GetMyHash function into BlockHeaderBase
