+++ src/libData/BlockData/BlockHeader/DSBlockHeader.h
@@ -63,10 +63,11 @@ public:
 
     /// Implements the Serialize function inherited from Serializable.
     unsigned int Serialize(std::vector<unsigned char>& dst,
-                           unsigned int offset) const;
+                           unsigned int offset) const override;
 
     /// Implements the Deserialize function inherited from Serializable.
-    int Deserialize(const std::vector<unsigned char>& src, unsigned int offset);
+    int Deserialize(const std::vector<unsigned char>& src,
+                    unsigned int offset) override;
 
     /// Returns the difficulty of the PoW puzzle.
     const uint8_t& GetDSDifficulty() const;
@@ -94,9 +95,6 @@ public:
 
     /// Returns the software version information used during creation of this block.
     const SWInfo& GetSWInfo() const;
-
-    /// Calculate my hash
-    BlockHash GetMyHash() const;
 
     /// Equality operator.
     bool operator==(const DSBlockHeader& header) const;
